review: round-2 fixes — gate plugin config eval, share host formatter, test CSP (#1795)
Address the re-review of #1796:

- A1: gate honoMiddlewarePlugin on isDevServer, not just its `apply`/the
  server.host ternary. Its config arg buildWebServerConfigFromEnv() (→
  resolveBindHostname) was evaluated eagerly, so an ambient HOST=0.0.0.0 still
  threw at config load for `vite build` and vitest. Comment corrected.
- A2: assert the full sandbox frame-ancestors directive (incl. http://[::1]:*)
  so the IPv6-loopback CSP fix is regression-guarded, not prefix-matched.
- A3: resolveBindHostname de-brackets its return (HOST=[::1] → ::1) so
  detection, listen(), and the origin list agree; formatHostForUrl re-adds.
- A4: drop the dead `value === ""` branch in isAllZeroIpv4.
- A5: `let webConfig: WebServerConfig` + match console.error("Error:", msg).
- A6: move formatHostForUrl to core/node/hostUrl.ts (shared) and apply it at
  the two remaining unbracketed sites — server.ts EADDRINUSE message and the
  CLI deep-link (cli.ts buildHandoff, which reads HOST). resolve-bind-host
  re-exports it; new core test covers it (100%).
